refactor: remove BridgeStopDetails and SwapStopDetails
