--- apps/cowswap-frontend/src/modules/bridge/pure/SwapStopDetails/styled.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components'
-
-export const WinningSolverContainer = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 4px;
-`
