+++ apps/cowswap-frontend/src/modules/bridge/pure/contents/BridgingProgressContent/FailedBridgingContent/index.tsx
@@ -3,7 +3,7 @@ import RefundIcon from '@cowprotocol/assets/cow-swap/icon-refund.svg'
 import { ConfirmDetailsItem, ReceiveAmountTitle } from 'modules/trade'
 
 import { AnimatedEllipsis, DangerText, InfoTextBold, InfoTextSpan } from '../../../../styles'
-import { StyledAnimatedTimelineRefundIcon } from '../../../BridgeStopDetails/styled'
+import { StyledAnimatedTimelineRefundIcon } from '../../styled'
 
 export function FailedBridgingContent() {
   return (
